+++ CHANGELOG.md
@@ -5,11 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.4] - 06-06-2023
+## [2.0.0] - 06-14-2023
 ### Changed
 
-- Updated the Relativity.OutsideIn package to 2023.4.0
+- Updated the Relativity.OutsideIn package to 2023.4.0 [REL-841948] (https://jira.kcura.com/browse/REL-841948)
 
+## [1.9.9] - 06-14-2023
+
+### Changed
+
+- Updated tagGitCommit logic to get server git tags [REL-850174] (https://jira.kcura.com/browse/REL-850174)
+
+## [1.17.1] - 06-08-2023
+
+### Changed
+
+- Aligned 'Import-API-RDC' versioning to trident standards [REL-850174] (https://jira.kcura.com/browse/REL-850174)
 
 ## [1.0.3] - 05-23-2023
 ### Changed